Enable neutron gw charm to control # of workers
Charm was missing this functionality and this patch is to add
this feature. We were trying to deploy this charm on bare metal
host machine with 72 cores. In end up creating 144 processes for
nova-metadata-api, and neutron.

This patch will provide the option to limit the number of worker
process in case of high number of cores in a host.

Enable it for icehouse, kilo, liberty and mitaka.

Partial-Bug: 1600482
Change-Id: Iedcbe6c53746a203937b78dd00e83fda7681bd8f