Switch to LBaaS v2 for OpenStack Newton
OpenStack Newton drops support for LBaaS v1; Switch to using the
v2 agent, with associate new template for driver options.

Upgrades from Mitaka will work as the old neutron-lbaas-agent
package will be removed as part of the dist-upgrade to the new
Newton packages and then the new neutron-lbaasv2-agent package
will be installed (as prescribed in the upstream migration
documentation).

Change-Id: I7ef26aca2f86f663366d463aeb3695ff5a402661