Correct package list post-upgrade
The list of packages to install post-upgrade was being retrieved
using the pre-upgrade OpenStack release version due to the
os_release function using a caches value.

This change syncs in charmhelpers which includes the reset_os_release
function to clear the cache. This charm now uses that function to
clear the cache after the dist-upgrade

Change-Id: I23f7af2b464c7e710423c15122c06bcea84e27a3
Closes-Bug: 1631897