newton: legacy external networking support
For OpenStack Newton, the external network bridge configuration
default switched from br-ex to '' inline with the preferred
method of configuring external networks using provider networks
via the l2 agent (instead of the l3 agent).

This had the side effect of breaking support for legacy style
networking.

Set:

  external_network_bridge = br-ex

in the event that new style external networking is not being used
to preserve backwards compatibility for now.

Note that this option will be removed in Ocata, so environments
need to upgrade to the preferred method of external networking in
order to stay supported.

Change-Id: Id929c0fa9baa0ecf0168d4d08f93939b56f847ab
Closes-Bug: 1634966