Add quantum-network-service as a mandatory service
The quantum-network-service relation with nova-cc needs to
be present otherwise metadata server has no keystone auth
credentials. This change adds it to REQUIRED_INTERFACES
so that its absence is reported in workload status. Also
the quantum-network-service-relation-broken symlink has
been added to trigger an assess status call when the
relation is removed.

Change-Id: I9859d532e1ef3c18c8c24308a115624035ef6d0f
Closes-Bug: 1634106