drop non-neutron tests
In order to avoid blocking testing of the neutron charm due to
non-neutron related failures, let's delete them from this charm's tests.

---

https://docs.openstack.org/releasenotes/nova/newton.html#deprecation-notes

"The /os-certificates API is deprecated, as well as the nova-cert
service which powers it."

e8f5341034c574aa07fb587e66c5c47a1d3620f8 blacklisted the package from
installation, however, the basic deployment test was not modified for
this charm which blocks CI for all changes.

Change-Id: Icf2ffd55d25e256ece35ef622d39112d27a25b61