Add QoS support
This patch adds support for reading the 'enable-qos' setting from the
neutron-plugin-api relation and adding 'qos' to the extension_drivers setting
if it is True. This is part of a wider set of changes to support QoS across
the neutron charms.

The amulet tests were missing the neutron-api to neutron-gateway relation this
has been added in. A side-effect of this is that the l2-population setting is
now properly being set to True so tests were updated to expect that.

A charmhelper sync was performed to pull in the QoS update to the
NeutronAPIContext.

Note: Amulet tests will fail until the corresponding neutron-api change
lands

Depends-On: I1beba9bebdb7766fd95d47bf13b6f4ad86e762b5
Change-Id: I6dc71a96b635600b7e528a9acdfd4dc0eded9259
Partial-Bug: #1705358